+++ x/group/keeper/msg_server_test.go
@@ -12,6 +12,9 @@ import (
 	"github.com/golang/mock/gomock"
 
 	"cosmossdk.io/core/header"
+	"cosmossdk.io/x/group"
+	"cosmossdk.io/x/group/internal/math"
+	"cosmossdk.io/x/group/keeper"
 
 	"github.com/cosmos/cosmos-sdk/codec/address"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
@@ -19,9 +22,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/cosmos/cosmos-sdk/x/group"
-	"github.com/cosmos/cosmos-sdk/x/group/internal/math"
-	"github.com/cosmos/cosmos-sdk/x/group/keeper"
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 )